+++ packages/api/src/detectPackage.ts
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { packageInfo as deriveInfo } from '@polkadot/api-derive/packageInfo';
-import { packageInfo as metaInfo } from '@polkadot/metadata/packageInfo';
 import { packageInfo as coreInfo } from '@polkadot/rpc-core/packageInfo';
 import { packageInfo as providerInfo } from '@polkadot/rpc-provider/packageInfo';
 import { packageInfo as typesInfo } from '@polkadot/types/packageInfo';
@@ -11,4 +10,4 @@ import { detectPackage } from '@polkadot/util';
 
 import { packageInfo } from './packageInfo';
 
-detectPackage(packageInfo, typeof __dirname !== 'undefined' && __dirname, [deriveInfo, metaInfo, coreInfo, providerInfo, typesInfo, knownInfo]);
+detectPackage(packageInfo, typeof __dirname !== 'undefined' && __dirname, [deriveInfo, coreInfo, providerInfo, typesInfo, knownInfo]);