+++ packages/metadata/src/util/index.ts
@@ -1,7 +1,0 @@
-// Copyright 2017-2021 @polkadot/metadata authors & contributors
-// SPDX-License-Identifier: Apache-2.0
-
-export { flattenUniq } from './flattenUniq';
-export { getUniqTypes } from './getUniqTypes';
-export { toCallsOnly } from './toCallsOnly';
-export { validateTypes } from './validateTypes';