+++ packages/types/src/generic/Block.spec.ts
@@ -4,11 +4,10 @@
 /* eslint-disable sort-keys */
 /* eslint-disable @typescript-eslint/no-unsafe-member-access */
 
-import { Metadata } from '@polkadot/metadata';
-import metadataStatic from '@polkadot/metadata/static';
-
 import { TypeRegistry } from '../create';
 import block00300 from '../json/SignedBlock.003.00.json';
+import { Metadata } from '../metadata';
+import metadataStatic from '../metadata/static';
 import { GenericBlock as Block } from './Block';
 
 const registry = new TypeRegistry();