+++ packages/types/src/metadata/types.ts
@@ -1,0 +1,9 @@
+// Copyright 2017-2021 @polkadot/types authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import type { Vec } from '../codec';
+import type { Codec } from '../types';
+
+export interface MetadataInterface<Modules extends Codec> extends Codec {
+  modules: Vec<Modules>;
+}