+++ packages/types/src/metadata/v12/Metadata.spec.ts
@@ -1,0 +1,17 @@
+// Copyright 2017-2021 @polkadot/types authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import { TypeRegistry } from '../../create';
+import { decodeLatestSubstrate, defaultValues, toLatest } from '../util/testUtil';
+import substrateData from './static';
+import substrateJson from './static-substrate.json';
+
+describe('MetadataV12 (substrate)', (): void => {
+  const registry = new TypeRegistry();
+
+  decodeLatestSubstrate(registry, 12, substrateData, substrateJson);
+
+  toLatest(registry, 12, substrateData);
+
+  defaultValues(registry, substrateData, true, true);
+});